requirements.txt
torchmetrics
pytorch_lightning
tensorboard
sentencepiece
protobuf==3.20.2
pytest
pybind11
lightning
stable_baselines3
deepspeed
litdata